Move SelectedFilesCard to the component package
