+++ app/src/main/java/com/infomaniak/swisstransfer/ui/screen/newtransfer/importfiles/components/SelectedFilesCard.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.swisstransfer.ui.screen.newtransfer.importfiles
+package com.infomaniak.swisstransfer.ui.screen.newtransfer.importfiles.components
 
 import android.content.Context
 import android.content.res.Configuration
